requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.3
diskcache==5.6.3
distro==1.9.0
docx==0.2.4
exceptiongroup==1.2.0
extra-streamlit-components==0.1.60
filelock==3.13.1
FlashRank==0.2.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.48
langchain-openai==0.1.4
langchain-text-splitters==0.0.1
langsmith==0.1.51
llama_cpp_python==0.2.67
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.17.3
openai==1.25.0
orjson==3.10.1
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
pyarrow==15.0.0
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pypdf==3.17.1
PyPDF2==3.0.1
pypdfium2==4.24.0
python-certifi-win32==1.6.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==69.5.1
setuptools-scm==8.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowflake-id==0.0.5
SQLAlchemy==2.0.25
streamlit==1.33.0
streamlit-authenticator==0.2.3
streamlit-chat==0.1.1
streamlit-pydantic==0.6.0
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
text-generation==0.6.1
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
transformers==4.37.2
triton==2.2.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==3.0.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
